Relax dependencies
We add `testImplementation()` dependency on Protobuf Gradle Plugin so that it's visible in the test Gradle project in `GeneratedDirPropertySpec`.

After we removed the `api()` dependency on the Protobuf Gradle Plugin adding it to a project requires either adding `buildScript` to the project with `gradlePluginPortal`. The hack with the plugin in the classpath works too.